+++ civis_jupyter_notebooks/assets/jupyter_notebook_config.py
@@ -25,6 +25,7 @@ c.NotebookApp.port = 8888
 c.NotebookApp.open_browser = False
 c.NotebookApp.token = ''
 c.NotebookApp.tornado_settings = {'headers': {'Content-Security-Policy': "frame-ancestors *"}}
+c.NotebookApp.terminado_settings = {'shell_command': ['bash']}
 c.MultiKernelManager.default_kernel_name = os.environ['DEFAULT_KERNEL']
 c.NotebookApp.allow_root = True