+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+- Improvements to terminal navigation
+- Default terminal to bash
+- Add civicon font
+
 ## [0.2.4] - 2017-11-28
 
 ### Fixed